Move "allItemsSize" variable inside "runExport" function
This commit relocates the "allItemsSize" variable from the global scope to the "runExport" function. It was unnecessarily polluting the global scope while serving a very specific purpose only used inside "runExport". It also adjusts how this variable is updated, passing a callback to the "exportForPage" function which adds the sizes accordingly. This architectural change improves data encapsulation and reduces potential interactions which might affect this variable accidentally.